+++ docs/examples/streaming/jsonl/auth.py
@@ -2,8 +2,8 @@ import dataclasses
 from collections.abc import AsyncIterator
 
 from dmr.plugins.msgspec import MsgspecSerializer
-from dmr.security.django_session import DjangoSessionAsyncAuth
-from dmr.streaming.sse import SSEController, SSEvent
+from dmr.security.jwt import JWTAsyncAuth
+from dmr.streaming.jsonl import JsonLinesController
 
 
 @dataclasses.dataclass(frozen=True, slots=True)
@@ -11,17 +11,14 @@ class _User:
     email: str
 
 
-class UserEventsController(SSEController[MsgspecSerializer]):
-    auth = (DjangoSessionAsyncAuth(),)
+class UserEventsController(JsonLinesController[MsgspecSerializer]):
+    auth = (JWTAsyncAuth(),)
 
-    async def get(self) -> AsyncIterator[SSEvent[_User]]:
+    async def get(self) -> AsyncIterator[_User]:
         return self.produce_user_events()
 
-    async def produce_user_events(self) -> AsyncIterator[SSEvent[_User]]:
-        yield SSEvent(
-            _User(email='first@example.com'),
-            event='user',
-        )
+    async def produce_user_events(self) -> AsyncIterator[_User]:
+        yield _User(email='first@example.com')
 
 
 # run: {"controller": "UserEventsController", "method": "get", "url": "/api/user/events/", "fail-with-body": false, "assert-error-text": "Not authenticated"}  # noqa: ERA001, E501